Add debug clicker
For now I cannot seem to find a parallel to Firefox unfortunately... :(

This is starting to get a bit janky, so maybe it would be best to go the route of sending requests over the network... I don't like the interface desync going that route, but it might be possible. We might have to figure out how to manually paginate through network requests as well.