Add tabs permissions
We'll need this for the debug clicker, but we'll also need to be calling this from a background worker since the tabs API isn't available in the content script.

I thought we were in the final stretch but I guess not!

Depending on how the messaging is implemented between the content script and background script/worker, we may not actually need the tab permissions if that's implicitly passed from the messaging protocol.